Match paper view count type (#40)

+++ user_views.go
@@ -16,7 +16,7 @@ type UserPaperViewRow struct {
 	PaperID      string
 	PaperTitle   string
 	Categories   string
-	ViewCount    int
+	ViewCount    int64
 	LastViewedAt time.Time
 }
